requirements.txt
langchain
langchain-openai
langchain-community
openai
chromadb
pypdf
python-dotenv
